requirements.txt
pillow
django
pandas
djangorestframework
markdown
django-cors-headers
requests